+++ tests/src/main/java/org/nervos/appchain/tests/TokenAccountExample.java
@@ -12,7 +12,7 @@ import org.nervos.appchain.protocol.core.methods.response.TransactionReceipt;
 
 public class TokenAccountExample {
 
-    private static int chainId;
+    private static BigInteger chainId;
     private static int version;
     private static String privateKey;
     private static String fromAddress;
